+++ atropos/commands/detect/cli.py
@@ -3,7 +3,7 @@
 from xphyle import STDOUT, STDERR
 from atropos.commands.cli import (
     BaseCommandParser, positive, readable_url, writeable_file,
-    readwriteable_file)
+    readwriteable_file, probability)
 
 class CommandParser(BaseCommandParser):
     name = 'detect'
@@ -79,6 +79,24 @@ you don't know what are the adapter sequences.
             action="store_false", dest="cache_adapters", default=True,
             help="Don't cache contaminant list as '.contaminants' in the "
                  "working directory.")
+        
+        group = self.add_group("Known Detector Options")
+        group.add_argument(
+            "--min-kmer-match-frac",
+            type=probability, default=0.5,
+            help="Minimum fraction of contaminant kmers that must be found "
+                 "in a read sequence to be considered a match.")
+        
+        group = self.add_group("Heuristic Detector Options")
+        group.add_argument(
+            "--min-frequency",
+            type=probability, default=0.001,
+            help="Minimum frequency required to retain a k-mer.")
+        group.add_argument(
+            "--min-contaminant-match-frac",
+            type=probability, default=0.9,
+            help="Minimum fraction of nucleotides that must align for a "
+                 "detected contaminant to match a known adapter sequence.")
         
         group = self.add_group("Output")
         group.add_argument(